Move proposal page to DAO, fix some issues
