+++ src/App.tsx
@@ -16,7 +16,6 @@ import DAOsPage from "./pages/DAOs"
 import Header from "./components/Header"
 import Footer from "./components/Footer"
 import ErrorPlaceholder from "./components/UI/ErrorPlaceholder"
-import Proposal from "./pages/Proposal"
 
 const AppWithEthers: FunctionComponent = () => {
 	const auth = useAuth()
@@ -34,7 +33,6 @@ const AppWithEthers: FunctionComponent = () => {
 						<Route exact path="/profile/:userId" component={Profile} />
 						<Route exact path="/dao/:address" component={DAOPage} />
 						<Route exact path="/daos" component={DAOsPage} />
-						<Route exact path="/proposal/:id" component={Proposal} />
 					</Switch>
 					<Footer />
 				</div>
