+++ src/components/Controls/Button/BackButton.tsx
@@ -2,13 +2,11 @@ import {FunctionComponent} from "react"
 import Button from "./"
 import {ReactComponent as ArrowDown} from "../../../assets/icons/arrow-down.svg"
 
-const BackButton: FunctionComponent<{onClick: () => void}> = ({onClick}) => {
-	return (
-		<Button onClick={onClick} buttonType="link" extraClassName="back-button">
-			<ArrowDown width="10px" height="20px" />
-			Back
-		</Button>
-	)
-}
+const BackButton: FunctionComponent<{onClick: () => void}> = ({onClick}) => (
+	<Button onClick={onClick} buttonType="link" extraClassName="back-button">
+		<ArrowDown width="10px" height="20px" />
+		Back
+	</Button>
+)
 
 export default BackButton
