+++ src/components/DAO/Proposal/VotesCard/styles.scss
@@ -1,4 +1,4 @@
-@import "src/variables";
+@import "../../../../variables";
 
 .votes-card {
   width: 280px;
