+++ src/components/DAO/Proposal/styles.scss
@@ -1,4 +1,4 @@
-@import "src/variables";
+@import "../../../variables";
 
 .proposal {
   display: flex;
