--- src/components/UI/Dashboard/styles.scss
@@ -1,11 +1,0 @@
-.dashboard {
-  width: 100%;
-  margin-top: 70px;
-  display: flex;
-  flex-wrap: wrap;
-  flex: 1;
-
-  &__content {
-    max-width: 880px;
-  }
-}
