+++ src/components/UI/Tag/index.tsx
@@ -2,9 +2,15 @@ import {FunctionComponent} from "react"
 import "./styles.scss"
 
 const Tag: FunctionComponent<{
-	variant: "active" | "executed" | "failed" | "canceled" | "executing" | "pending" | "queued"
-}> = ({children, variant}) => {
-	return <span className={`tag tag--${variant}`}>{children}</span>
-}
+	variant:
+		| "active"
+		| "executed"
+		| "failed"
+		| "canceled"
+		| "executing"
+		| "pending"
+		| "queued"
+		| "outdated"
+}> = ({children, variant}) => <span className={`tag tag--${variant}`}>{children}</span>
 
 export default Tag
