+++ src/hooks/getters/useMyDAOs.ts
@@ -31,7 +31,7 @@ const useMyDAOs = (): {
 						const firebaseData = await getDAO(dao)
 						const gnosisVotingThreshold = await getVotingThreshold(dao, provider)
 						const owners = await getOwners(dao, provider)
-						const strategies = await getStrategies(dao, provider)
+						const strategies = firebaseData.seeleAddress ? await getStrategies(dao, provider) : []
 						const tokenSymbol = ""
 						const balance = 0
 						const fundedProjects = 0
