+++ src/types/safeProposal.ts
@@ -18,7 +18,7 @@ export const SafeProposalsTypeNames = {
 	decentralizeDAO: "Decentralize DAO"
 } as const
 
-export type SafeProposalState = "active" | "executed" | "canceled"
+export type SafeProposalState = "active" | "executed" | "outdated"
 
 export type SafeProposal = {
 	id?: number
